+++ src/encoding/evm/strategy_encoder/mod.rs
@@ -1,3 +1,4 @@
 pub mod strategy_encoder_registry;
 mod strategy_encoders;
 mod strategy_validators;
+mod utils;
